Source approach
In this approach we don't build the libraries up front, we just use the
src files directly and only compile the entire thing in the
`main-package`. This generates all the chunks that we need.

I have noticed that it's a good idea to actually prefix each
"chunk-name" with something, e.g. the package name, so it's easier to
identify and group them.